+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## [0.2.1] - 2019-11-24
 
 ### Added
-- Extend / pretty-print JavaScript-Errors with [`stacktracey`](https://www.npmjs.com/package/stacktracey) 
+
+- Extend / pretty-print JavaScript-Errors with [`stacktracey`]
 - Ability to show last 5 values per Watcher
+
+[`stacktracey`]: https://www.npmjs.com/package/stacktracey
 
 ## [0.2.0] - 2019-10-28